backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
Flask==2.3.2
Flask-CORS==3.0.10
gunicorn==20.1.0
psycopg2-binary==2.9.7
bcrypt==4.0.1